Extract collections benchmarks to libcollections/benches And libcore/benches
