+++ src/libcore/benches/hash/mod.rs
