+++ src/libcore/benches/hash/sip.rs
