+++ src/libcore/benches/lib.rs
