+++ src/libcore/benches/mem.rs
