+++ src/libcore/benches/ops.rs
